// src/app/generated/model/taskResponseInner.ts
/**
 * TriReg2 API
 * An OpenAPI specification for time registrations
 *
 * The version of the OpenAPI document: 0.0.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TaskResponseInner { 
    taskId: string;
    /**
     * The name of the task
     */
    name: string;
}