// src/app/generated/model/timeRegistrationResponse.ts
/**
 * TriReg2 API
 * An OpenAPI specification for time registrations
 *
 * The version of the OpenAPI document: 0.0.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeRegistrationResponse { 
    registrationId: string;
    /**
     * The registration time as Unix milliseconds timestamp
     */
    registered: number;
    userId: string;
    taskId: string;
    date: string;
    /**
     * ISO 8601 formatted duration component <PnDTnHnMn.nS>. Each unit is optional, but at least one must be specified
     */
    duration: string;
    status: string;
}